Change phi exponentiation
In order to make the phi factor more controllable, it is interesting that
cities equally spaced from a reference point are complementarily favored.
To achieve this behavior, the phi vector, as it is currently implemented,
needed to be elevated to q squared, instead of only q. When plotted in a
log scale, phi is now linear.